chore: initial commit with working UDP
The rest of the PoC is unfinished and I doubt I will finish it
anytime soon; however, the interaction between the network layer and
the conn layer is quite interesting[*] and I'd like to keep it.

[*] I mean, there are probably many other Go projects using this
pattern around, but this is the first time this kind of pattern clicked
for me, so I would like to keep these files around.